Add more valid move checks
Add the move checks for all pieces but the pawn.
Add a function for checking if a spaces is a valid space (on the board).
Change all storages of sets of spaces or moves to using Python sets.
Make sure all move checks will return false if the start space is the
same as the end space.